Fully import Zip Linq tests from scripted legacy import
Converts the Zip tests imported in 2874 into a more directly XUnit-ian style.
